Time: 5 ms (99.22%), Space: 60.5 MB (40.08%) - LeetHub

+++ 1339-maximum-product-of-splitted-binary-tree/1339-maximum-product-of-splitted-binary-tree.java
@@ -17,7 +17,7 @@ class Solution {
 
     public long answer;
 
-    public int dfsMax(TreeNode curr) {
+    public long dfsMax(TreeNode curr) {
 
         if (curr == null) {
             return 0;
@@ -39,7 +39,7 @@ class Solution {
 
     public int maxProduct(TreeNode root) {
 
-        int max = dfsMax(root);
+        long max = dfsMax(root);
         dfsAnswer(root, max);
 
         return (int) (this.answer % 1_000_000_007);
